+++ packages/client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Avoid throwing when creating a transaction that has an unknown signed-extension.
+- Validate transactions on Runtimes that only support metadata V14.
 
 ## 1.4.1 - 2024-10-05